fix ninja build

+++ tasks/seq/tsatsyn_a_vector_dot_product/src/ops_seq.cpp
@@ -7,7 +7,7 @@ using namespace std::chrono_literals;
 
 int tsatsyn_a_vector_dot_product_seq::resulting(const std::vector<int>& v1, const std::vector<int>& v2) {
   int res = 0;
-  for (int i = 0; i < v1.size(); ++i) {
+  for (size_t i = 0; i < v1.size(); ++i) {
     res += v1[i] * v2[i];
   }
   return res;
@@ -35,9 +35,7 @@ bool tsatsyn_a_vector_dot_product_seq::TestTaskSequential::validation() {
 
 bool tsatsyn_a_vector_dot_product_seq::TestTaskSequential::run() {
   internal_order_test();
-  for (int i = 0; i < v1.size(); i++) res += v1[i] * v2[i];
-  // std::this_thread::sleep_for(20ms);
-
+  for (size_t i = 0; i < v1.size(); i++) res += v1[i] * v2[i];
   return true;
 }
